- Removed IBlitableSurface and folded it back into Surface. This got rid of some ugly casting. - get_w, get_h -> width, height in some places - surfaces and pictures height and width are not uint16_t only. - Converted some max,min calls from using casts to explicit template types.

+++ src/graphic/image_cache.h
@@ -61,8 +61,8 @@ public:
 	/// error.
 	// NOCOM(#sirver): document this
 	virtual const IPicture* get(const std::string& hash, bool alpha = true) = 0;
-	virtual const IPicture* render_text(const std::string& text, uint32_t w) = 0;
-	virtual const IPicture* resize(const IPicture*, uint32_t w, uint32_t h) = 0;
+	virtual const IPicture* render_text(const std::string& text, uint16_t w) = 0;
+	virtual const IPicture* resize(const IPicture*, uint16_t w, uint16_t h) = 0;
 	virtual const IPicture* gray_out(const IPicture*) = 0;
 	virtual const IPicture* change_luminosity(const IPicture*, float factor, bool halve_alpha) = 0;
 };
